feat: Componentizando o button
