+++ app/lib/global/app_text_style.dart
@@ -5,4 +5,9 @@ class AppTextStyle {
 
   static const TextStyle headerText =
       TextStyle(fontSize: 24, fontWeight: FontWeight.bold);
+
+  static const TextStyle mediumText =
+      TextStyle(fontSize: 16, fontWeight: FontWeight.bold);
+
+  static const TextStyle smallText = TextStyle(fontSize: 14);
 }
