+++ app/lib/layout/lateralMenu.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:invenmanager/global/app_color.dart';
+import 'package:invenmanager/global/app_text_style.dart';
 import 'package:invenmanager/pages/inventory_history_page.dart';
 import 'package:invenmanager/pages/profile_page.dart';
 
@@ -9,7 +10,7 @@ class LateralMenu extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Drawer(
-      backgroundColor: AppColor.black,
+      backgroundColor: AppColor.gray_900,
       child: ListView(
         padding: EdgeInsets.zero,
         children: <Widget>[
@@ -22,12 +23,10 @@ class LateralMenu extends StatelessWidget {
                   radius: 30,
                   backgroundImage: AssetImage('lib/assets/profile_image.jpg'),
                 ),
-                const Text(
+                Text(
                   'John Doe',
-                  style: TextStyle(
-                      color: AppColor.white,
-                      fontSize: 18,
-                      fontWeight: FontWeight.bold),
+                  style:
+                      AppTextStyle.mediumText.copyWith(color: AppColor.white),
                 ),
                 TextButton(
                   onPressed: () {
@@ -36,9 +35,10 @@ class LateralMenu extends StatelessWidget {
                         MaterialPageRoute(
                             builder: (context) => const Profilepage()));
                   },
-                  child: const Text(
+                  child: Text(
                     'Ver perfil',
-                    style: TextStyle(color: AppColor.yellow),
+                    style:
+                        AppTextStyle.smallText.copyWith(color: AppColor.yellow),
                   ),
                 ),
               ],
@@ -48,9 +48,10 @@ class LateralMenu extends StatelessWidget {
             crossAxisAlignment: CrossAxisAlignment.center,
             children: [
               ListTile(
-                title: const Text(
+                title: Text(
                   '85 Produtos cadastrados',
-                  style: TextStyle(color: AppColor.gray_200),
+                  style: AppTextStyle.mediumText
+                      .copyWith(color: AppColor.gray_200),
                 ),
                 onTap: () {
                   Navigator.push(
@@ -60,9 +61,10 @@ class LateralMenu extends StatelessWidget {
                 },
               ),
               ListTile(
-                title: const Text(
+                title: Text(
                   '9 produtos em falta',
-                  style: TextStyle(color: AppColor.gray_200),
+                  style: AppTextStyle.mediumText
+                      .copyWith(color: AppColor.gray_200),
                 ),
                 onTap: () {
                   Navigator.push(
@@ -80,9 +82,9 @@ class LateralMenu extends StatelessWidget {
           ),
           ListTile(
             leading: const Icon(Icons.exit_to_app, color: AppColor.yellow),
-            title: const Text(
+            title: Text(
               'Sair',
-              style: TextStyle(color: AppColor.yellow),
+              style: AppTextStyle.mediumText.copyWith(color: AppColor.yellow),
             ),
             onTap: () {},
           ),
