+++ app/lib/pages/recover_password_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:invenmanager/global/app_color.dart';
-import 'package:invenmanager/pages/login_page.dart';
+import 'package:invenmanager/pages/login/login_page.dart';
 
 class RecoverPassword extends StatefulWidget {
   const RecoverPassword({Key? key}) : super(key: key);
